Replaced `Interlocked.{Add, Decrement}` with `InterlockedAddWithOverflowCheck`. Added `Retain`/`MaybeDispose` around `Race.Handle` to try to fix rare test crashes. Exit lock sooner in `AsyncPromiseRef` progress subscribe.

+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/HelperFunctionsInternal.cs
@@ -432,7 +432,7 @@ namespace Proto.Promises
         }
 
         [MethodImpl(InlineOption)]
-        private static long InterlockedAddWithOverflowCheck(ref int location, int value, int comparand)
+        private static int InterlockedAddWithOverflowCheck(ref int location, int value, int comparand)
         {
 #if PROMISE_DEBUG || PROTO_PROMISE_DEVELOPER_MODE
             Thread.MemoryBarrier();
